Rework the PAM module logic for new API and back behaviour
We now allow from the PAM module:
- to skip one or many steps, depending on the last broker used for a
  given user and last authentication used.
- go back one or multiple steps when entering 'r'.
- reduce the number of user interactions in general.

The return logic will use 'escape' later on, but it needs basing the UI
in a TUI to work in every circumstances, including wait only scenarios.

We still need the specific register last broker used for an user, once
we know that the authentication logic passed (even for local user).

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>