Set default user for broker once the user authentication is valid
In acct management, we know the user credentials are valid. So it exists
against that broker.
We need to ensure that local broker usage is "saved" too. For this, we
keep the sessionID between calls and find back the corresponding broker
for this session ID to save it for the user.

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>