Validate UI layout based on PAM specification
We used to have a static verification for the UI layouts returned by the
broker, but those are subject to change according to what the system can
currently handle.

This means that it's up to PAM to provide which layouts are supported,
which fields of those layouts are required and what values those fields
support.

This replaces the previous static implementation of validateUILayouts to
a dinamic one that takes into consideration what was specified by PAM.