pam: Add GDM JSON protocol definition
We use JSON to communicate with GDM since its UI, GNOME Shell is in
JavaScript and so can handle those messages natively pretty well.

Here we define the primitives of this protocol and their conversion back
and forth from go structures and types, ensuring the expectations are
met.

The protocol code generation is mostly done via protobuf so that we can
avoid manual definition of many structures, but is meant for being
used from the JSON side too thus using some naming conventions to please
both cases.