Broker configuration path as part of constructing brokers manager
This is a mandatory argument for brokers manager as we base our detected
broker configuration on this. Remove the options then, and rebase all
tests to include it.
Also, standardize for this argument on path vs dir, as this is really a
path in the end.
Fix a configuration stutter in pam_test.go, as we wanted to really
return the brokers configuration path and not the temporary base
directory.