pam: Use pam-go to generate module stubs and handle PAM APIs
Avoid doing all the C manual work inside authd pam, but handle this in a
separate library that handles this all and comes with fully tested
operations.

We could so mock the ModuleTransaction here if we want so that we can
make things more testable as they are now.

As per this change, the module can be simply be generated with:

  go generate -C pam