Integration tests for NSS
Adding the integration tests for the NSS library to cover its
interaction with authd. Due to the new features introduced in Go 1.20,
we can now build Go binaries with instrumentation enabled, which means
we can run the tests as they were supposed to (by building the daemon
binary and running it as a helper process).