pam/gdm/extension: Add PAM gdm JSON extension protocol implementation
Add go bindings to encode and decode PAM binary messages using the GDM
Custom JSON protocol.

The protocol will be available as part of GNOME 46, but for being early
consumers we include in the repo the headers where the messages are
defined, until that's possible to switch on the repo version.

See: https://gitlab.gnome.org/GNOME/gdm/-/merge_requests/227