Fix test script to prevent error without Rust cov
Sometimes, we want to run only specific tests that don't include the NSS
module (which means no Rust coverage). The script was quitting with an
error if it could not merge the Rust coverage to the Go one. This is
fine for the CI, since we want all coverage, but the script needs a
change to prevent this.