Fix authd daemon hanging on shutdown
NSS clients keep the library loaded until the end of the session, e.g.
if one runs a command like `sudo sleep 10`, the library won't be closed
until the sleep finishes. This also applies to SSH and login sessions -
easy to monitor by applying a inotify watch on the NSS library path.

One side effect of this was that due to the way we were initializing our
tokio runtime, gRPC clients would be left hanging until the end of the
aforementioned session, terminating together with the runtime. We
assumed that the gRPC connection would go out of scope and terminate
after the NSS method returns, but it appears this was not the case.

Leaving clients hanging meant that as long as there was at least one
open session (be it either a logged in user, a running sudo command, or
any other blocking third-party program interacting with NSS), the authd
daemon wouldn't be able to gracefully stop in a quick manner, thus
hanging the process and ultimately timing out after a period of time.

Because tonic (our Rust gRPC library) doesn't provide a bespoke way of
closing the client connection, we still need to rely on Rust to handle
it for us, so the proposed way of doing it is to declare the tokio
runtime inside the function call, which will make it go out of scope and
terminate the client afer it returns. We initialize the runtime the same
way, ensuring we remain single-threaded by using the current-thread
scheduler.