Increase wait time before exiting function
This is causing flaky tests in error cases where the test DB is corrupt
or has bad entries. I've tried to increase sleep times within the tests
but it appears the issue is in the implementation itself.

The function exits if it's not able to send something on the doClear
channel within 1 millisecond -- given that a cleanup might already be
running this would exit without queueing any subsequent cleanup, which
in turn left us with a dangling "dirty" DB file.

I think the ideal fix here would be to make this a buffered channel and
be able to queue multiple cleanups while one is running, but for now
I've increased the amount of time we wait before exiting.

Fixes UDENG-1847