Continue if active UID is not a real user
In chroot environments where /proc is mounted it's common to have
processes owned by UIDs that do not exist within the chroot.

I think getting the active users list should be a best effort approach
where we don't fail hard if we cannot do this mapping for an user,
mainly because it's highly likely for that user to not be in the DB we
want to clean up anyway.