Make group tests infrastructure reusable
We will need it in other places for integration tests. Make it then
exportable on its own package, using the linkname to change set
global defaults as we don’t want to leak the With* tests helpers
throughout the stack.

However, the tests using those won’t be able to be run in parallel.