Ensure that we differentiate gpasswd being called or not in tests
Now, we are going to not pre-create gpasswd output file but only create
it once the mock is called. That allows us to separate it being called
or not.
This functionality will be better in testutils, but it needs to all
post-treatment for the file to be idemnpotent and I didn't find an
elegant way for it.

Then, we can replace the LoadWithUpdateFromGolden with something
treating files (or maybe take back the tree comparison).