pam: Depend on go-pam v2
Upstream decided to switch to a v2 major version, since we included some
major changes in the application-side API, this implies changing the
imports as per Go policies.

Thus, I've also rebased my fork on upstream master, so that we don't
have to do this kind of refactor again soon.