pam/gdm: Support conversations using JSON data sent to PAM via gdm binary protocol
Add an utility function to perform GDM-PAM binary conversations using
the custom JSON protocol.

This function does not much per se, but it allows us to do some proper
testing via the dummy module transaction and so we can test how this
data is passed back and forth the two sides, and stress the C <-> go
structures conversions.