Skip pam-moduler call if dependencies are vendored
We depend on github.com/msteinert/pam/cmd/pam-moduler to generate the
pam_module.go file for the PAM module. This poses an issue if we are
trying to run it with vendored dependencies, as we cannot vendor the cmd
package.

This can be fixed multiple ways, but in the interest of keeping the fix
close to the problem, and avoid moving the generate statement to a
separate package/file, simply skip the call if we detect a vendor
directory at the root of the project.