Disallow brokers from returning users without groups
The brokers used to be able to provide the userinfo without any group
associated to it. This creates difficulties when trying to manage the
user presence in the local groups. Now, the broker must provide at least
one group for the user and the first group must be a remote one (with
UGID).