Bump the toolchain to fix vulnerabilities
Set to toolchain to 1.21.5 to avoid some vulnerabilities in our code
path.
Downgrade the go mod version itself as in term of language, we are
compatible with it and there is no bug fixes requiring a newer version.
The desired toolchain can still be overriden if desired, but it helps
knowing our requirements for fixing those vulnerabilitie issues.