Give explicit PR permissions to GitHub token
When dependabot triggers a PR the GITHUB_TOKEN is read-only and must
explicitly be given write permissions.[1] This is a side effect of
upgrading to the v5 create-pull-request-action.

Explicitly set these in the workflow to avoid it failing on commits by
dependabot, but also to more clearly define the workflow's expected
permissions.

[1] https://github.com/peter-evans/create-pull-request/issues/1873