generate: Simplify generation inside go files, moving into a script
We need to reuse the proto generation code in gdm side too, so let's
cleanup it and move under a different script that is easier to reuse.

Also add the generate build tag to the go file so that is just ignored
at build time as it's supposed to be.