pam/go-loader: Allow providing an AUTHD_PAM_MODULES_PATH base dir
When the pam module is loaded without a full path it can't be loaded by
our dynamic loader since it doesn't know where it's supposed to be.

As per this, make it a requirement to have AUTHD_PAM_MODULES_PATH
defined, we provide a fallback only on the standard case, but the build
system should define it for all.