pam: Introduce pam_debug build tag to generate debug libraries
When using this tag on `go build` or `go generate` we:
 - Add more debug symbols to built libraries
 - Enable pedantic checks on data structures used by GDM proto

To achieve this, add a further generate_debug.go file that is only
used when the `pam_debug` tag is used, this now also builds the gdm
code with the `pam_gdm_debug` build tag that implies enabling by
default some checks that are not enabled by default.

Update CI job to ensure this works.