Skip NSS lookups from dbus-daemon through systemd
While doing more tests that required frequent reboots I still
encountered the boot hang issue described in UDENG-1862. I was able to
trace it back to an issue in systemd where something similar happened
between nss-systemd and dbus[1].

Even with the strictest systemd unit dependency ordering we're still not
able to rule out the deadlock completely, because while dbus.service may
report that it's active, other services could still interact with it
afterwards causing it to perform more _blocking_ NSS lookups, before the
authd daemon has fully started.

To prevent this from happening, we can leverage the systemd fix[2] which
sets a specific environment variable when starting the dbus.service,
telling itself (nss-systemd) to avoid using dbus for the lookup.

It's not really ideal to return an error here but it was the quickest
way to do it without changing all the NSS entry points of the library.

[1] https://github.com/systemd/systemd/issues/15316
[2] https://github.com/systemd/systemd/pull/22552