Add functions to clean the local groups
We already had a function to update the groups, but none to clean the
groups from inactive users or to just remove the user from all the
groups it belongs to.

Now we have one to clean a specific user and one to clean all inactive
users from the groups, based on the list of users returned by the
getpwent call.