Add newusers as transitional pkg for users
We realized that having two separate packages controlling user
information would complicate a lot of things in the long term. So,
instead, we are moving both the cache and the localgroups logic to the
users package and it will become a user manager for it.

This means that we need to move all "entrypoint" functions to it and
then map them to their respective packages (such as cache or localgroups
[when it comes])