pam: Add PamClientType enum type and use to identify clients
Our model may handle multiple pam client types, and depending on them it
could behave differently, so indicate this through an enum instead of a
simple boolean, since we will have multiple types not just interactive
and non interactive clients.

As for now, just return an error if a non-interactive client is used,
since that's the only thing we support right now.