Add flag to cover all packages
By default, Go only calculates test coverage of the package being
tested. While this is good to see how effective our unit tests are,
we also miss the coverage on more complex tests that test more than
one package at a time and this is specially harmful when dealing with
subpackages.