Implement encryption in examplebroker
This generates an asymmetric private/public key pair at broker creation
time. We let it panic as this broker is only for testing purposes.

This encryption key is then passed at session creation. Note that the
broker can create a new key for each session if desired.

The chosen encryption is sha512. The public key is passed as X.509 standard,
encoded as base64.

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>