Implement pam-side encryption mechanism
Follow the spec to encrypt any given "challenge" to follow the
encryption mechanism.

This is using the public key received at session creationg time. Note that
the broker can create a new key for each session if desired.

The chosen encryption is sha512. The public key is passed as X.509 standard,
encoded as base64.

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>