pam/integration-tests: Add GDM module integration tests
Compile the module as a PAM library and load it using go-pam via libpam
itself, at this point we use the defined conversations to communicate to
the module using the JSON protocol, ensuring that the expectations are
met.

This ensures that things work fine when we're in a real scenario.