Remove brokers subpkg from testutils
Having a subpackage inside testutils is a bad thing as it gets way more
confusing to navigate our test suite and tests in general. Now that we
don't rely on internal/ anymore, there's no issue with import cycles and
the broker utilities can be moved back to the testutils package.