systemd, debian: Install authd in /usr/libexec
authd is not a program that is meant to be ran by user, so there's no
point for it being in PATH, as it's systemd running it.

So move it where most daemons stay these daysystemd, debian: Install
authd in /usr/libexec

authd is not a program that is meant to be ran by user, so there's no
point for it being in PATH, as it's systemd running it.

So move it where most daemons stay these days