pam/adapter: Do not resize brokwerselection window on window changes
In terminal mode we're monitoring the window size and eventually changing
the broker selection list (only) that forces a fullscreen view of them.

This is something that it's not needed on the kind of "app" we're
building because it brings user context away (e.g. What was I
authenticating for?!) but also it makes testing the interface very hard
because will make bubbletea to potentially hide some error messages.