Tune NSS logging to Info instead of Debug
When using the Debug level, we were getting spammed by a lot of gRPC
connection logs from tonic and it looks like we can't control the
logging level individually. So, instead, let's tune down our log
messages to Info instead of Debug so that we can have a cleaner output
of what we want.