Set default group for user
We used to rely entirely on the groups information provided by the
broker, but this is unreasonable since we have some default groups in
unix systems that are "required" but only make sense in the OS context.

To prevent errors and reduce frustration from misconfiguration, we now
ensure that the user will always be a part of his own group and that it
will be the main group as well.

Co-authored-by: Gabriel Nagy <GabrielNagy@users.noreply.github.com>
Co-authored-by: Didier Roche-Tolomelli <didrocks@ubuntu.com>