authd will now wait for examplebroker to start
If built with the "withexamplebroker" tag, we expect authd to configure
and run the examplebroker as a "sub" service. We used to delegate the
starting process to a go routine, but this was creating problems in some
test environments (as the examplebroker was taking too long to start).
Now, starting the examplebroker is a blocking call and the daemon will
wait until the broker is up and running.