Add timeout when connecting to authd socket in PAM
This call could hang forever if the file was present but not ready yet,
so we should add a timeout instead to prevent deadlocks.

The documentation states that "By default, it's a non-blocking dial
(the function won't wait for connections to be established, and
connecting happens in the background). To make it a blocking dial,
use WithBlock() dial option. In the non-blocking case, the ctx does
not act against the connection. It only controls the setup steps.
In the blocking case, ctx can be used to cancel or expire the
ending connection."

This means that we need a blocking Dial if we expect the timeout to
happen "automatically" when the context gets canceled.