pam: Add an exec-client that runs the pam module using the dbus transaction
This allows us to have a binary file that can be launched by the exec
client and that can communicate with it.

This is now the new main that can be built without any tag, since it's
something we'll need to have in the default install.

As per this change we need to update the pam module to use values that
can be marshalled as variants for data values, this is not really a
problem for now.