Allow GetPreviousBroker to return "local"
When we can’t get an user from cache but can still resolve it through
NSS, it means that another service than authd is responsible for it
(passwd, winbind, sss…). We then return "local" and store it in memory
to avoid further NSS lookups for this user.
We do not store it in our database to avoid syncing this list of local
user with the ones existing on the system.