Authorizer to handle unix socket peer user
The Authorizer can decode from a context then current per user id (and
pid) which connected to an unix socket.
Then, it provides methods for authorization. One of them is only
authorizing if the current user is root.

Allow, with a private option, to override which uid is considered as
"root".