pam/go-exec: Use simple waitpid to monitor children lifetime
Instead of relying on signal as GLib does just use waitpid on a separate
thread as this is safer to use because our library might be loaded by
other applications that may set a different handler on SIGCHLD, breaking
us, and making our wait to be never-ending.

This is particularly the case of sshd, because it creates a subprocess
where actions are executed, replacing our SIGCHLD handler causing the
module to hang forever during a second call to setcred.