pam/go-exec: Optimize the wrapper module not to do unneeded ops
There are various operations that we have not implemented in our PAM
implementation in Go, so there's no need to make it more complex the C
side either.

So keep them as simple as possible, by just returning the same that we'd
return in the Go side. Adding tests to ensure that if something need
changes, we can track it better.