Store the session username and check returned one
Now the broker will remember the username provided when creating a new
session and use it to validate the one that is returned by the broker.

This way, we can prevent a case where the user selects a different
username than the one of the account he used to authenticate remotely.