Applying similar rule to localgroups
Split and rename the test helper functions, remove them from the
codebase.
Do not rely on build tags inside the code itself, this allows to build
some parts of the binary not having to rely on env variables to be
present here.
Centralize the env variable location. We will rename them in a separate PR.