Enable mkhomedir module via the PAM configuration
Ensure that home directories are automatically created for
authenticating users.

Because this is closely coupled to the authd functionality, I've opted
for enabling this via our PAM configuration as opposed to using
pam-auth-update. By doing this we also benefit from the cleanup that
would've been impossible had we used pam-auth-update.

Fixes #370 / UDENG-3117