Rework newpasswd view for CLI
Instead of showing both prompts from the get-go, and having Enter submit
the form straight away (and only being able to navigate through it with
Tab), such as:

    Enter your new password
    > ****
    >

We now render a single input if the user has just been shown the form,
with an updated prompt:

    Enter your new password

    New password:
    > ****

When pressing Enter -- provided that the password passes our quality
checks -- the form changes to:

    Enter your new password (confirm)

    New password:
    > ****
    Confirm password:
    > ****

This makes it clearer that we want the user to re-type their password.
As a result, Tab navigation is no longer supported in non-skippable
views, and the form can only be advanced by pressing Enter.

For skippable views (i.e. with a button), Tab only works if the password
is not filled in. This avoids opening a can of worms regarding to what
fields we should be showing or not based on whether password fields are
filled in. Thus, the skippable experience becomes:

    Enter your new password (3 days until mandatory)

    New password:
    >

    [ Skip ]  // Focusable

and

    Enter your new password (3 days until mandatory)

    New password:
    > *****

    [ Skip ]  // No longer focusable, have to finish filling in the form
              // (or clear it)

Otherwise, I believe we should keep the form as simple as possible,
without the possibility of jumping back to the previous password field
once it was filled in, at the sake of improved maintainability.

Fixes #365 / UDENG-3112