examplebroker: Simulate regeneration of qrcode for real
Example broker did support qrcode's code regeneration, but this didn't
lead to new content or code, making testing its actual functionality
harder.

So, each time the qrcode is selected we now loop through different (but
predictable) URIs and increase the code value.

As per this regenerate golden files so that the regeneration is tracked

Add support for `user-integration-static-qrcode` users so that we don't
have to track the content changes in brute-force tests where we don't
have a clear number of regeneration events and we don't care about such
changes.