pam/authentication: Start new auth only after cancellation is completed
In case we're still in the process of cancelling the authentication we
should not continue with the current action until we're notified that
the cancellation is completed.

This is needed because otherwise we may end up into this situation:
 - reselectAuthMode
 - AuthModeSelected
 - ...
 - startAuthentication
 - isAuthenticatedResultReceived: cancelled

Or a bit less extreme, but still getting the cancelled event while we're
selecting the authentication mode. And this mis-order causes troubles
especially when passing the data to gdm.

To avoid this, add a tracker to wait for the authentication cancellation
to be completed and only once we've received a response we can proceed
with next actions.

Due to the blocking nature of channels we need to use more async
functions now and add a few synchronization points

In particular we need to add some sleeps to workaround two main issues:
 1. Wait for the IsAuthenticated to be delivered before cancelling it
 2. Wait to start IsAuthentication requests after cancellation error

 # 1. Sleep between the request and its cancellation

When quickly calling IsAuthenticated followed by context cancellation,
we may end up in a situation in which we cancel the request before we've
actually placed one on the broker.

This is due to the fact that the dbus request may take some time to be
delivered and being stored by a remote broker, while the cancellation
may happen very quickly.

So, to prevent this, in case the context has been cancelled, let's wait
"a bit" before sending the cancellation request, so that we can be sure
that the `IsAuthenticate` request has been delivered before.

This is far from being nice, but it's the best way to achieve this
result given the API constraints that we have, and the fact that go-dbus
can't advertise us if a message has been delivered, without waiting for
the whole result.

See example log of this failure at: https://www.pastery.net/afkmuy/

 # 2. Sleep after receiving the cancelled error

The IsAuthentication request on client side may happen even while we're
still cancelling the previous one, this is because the client receives
the cancellation error as just after the calling context is cancelled,
and not when the server side has actually completed and returned the
request. This is due to the nature of the grpc.ClientConnInterface
used, and there's no simple way to have this information client-side
without rewriting the client low level implementation.

It's very easy to reproduce this case by just increasing the sleeping
time we added to fix previous point, making the cancellation to be
slower than actually is.

To workaround this, wait some reasonable extra time before delivering
the cancelled event, so that we can be confident enough that the server
side has received and cancelled/completed the previous request.

This is far from optimal, but we have no other way at client side to
know that the server has finished handling the all the previous requests.