Download NSS lib errors to info
That way, we don’t spam logs for 3rd party applications using our
library as they are the one which should control the output to the user.
We send back those errors in a NSS format anyway to the caller already.

We thus only print details when AUTHD_NSS_INFO=stderr is set.