pam/authentication: Store the previous challenge on success
We used to store the challenge early and unset it in case of
authentication failures, but this won't work well with the model that
GDM will use, since it will rely on the authentication result.

But it's also unsafer with other models, since we may leave stored
something that has not been acked, so only store the previous value
if it has been accepted by the broker.