examplebroker: Require different password if current password has been updated
When trying to change the password again when it has been changed to
`authd2404`, we'd be now blocked by the password quality check since
the new password is matching the previous one, so in case it has been
changed already, reset it to `goodpass` again.

As per this also spot an issue we had in the native tape that didn't
really check what we wanted.