pam/authentication: Ensure cancellations are forwarded
When triggering quick authentication re-selection requests we may end up
overwriting the local cancelAuthFunc while another one has is in process
of being set.

Due to the nature of bubbletea, saving the data in the model isn't fully
reliable since the model may not be updated when another request is
coming concurrently.

So, use the cancel function as the main authTracker check point so that
it's always the same during each authentication request, and it allows
to simplify some codepaths.

This is very easy to trigger by using the CLI, by just holding the Enter
key for a while, and the regeneration will stop, but it's not something
I've been able to test in VHS since holding a key isn't really possible
and the minimum key press delay isn't enough to trigger it.