Move generatedID to users and generate ID based on username
Relying on the broker name + UUID meant that the user would always have
to authenticate first before getting a uid. This is fine for cases where
the session is refreshed and updated constantly but, in others (such as
SSH), the user information is cached and used throught the
authentication+session, so the id was never refreshed and problems
happened.