pam: Wait for bubbletea to be done before resetting/closing the logging
bubbletea's Run() returns as soon as it receives a quit message, however
there may still be goroutines running that have not been completed and
that may deliver messages with some delay.

To prevent these to cause us to log too early to stderr, and breaking
our golden files, let's wait a bit before closing the logging.