debian: Also insert authd as NSS source on reinstall
The debian/postrm script removes authd as a source from
/etc/nsswitch.conf, but the debian/postinst script only added it when
the package was installed for the fist time. That led to NSS not using
authd as a source when the package was reinstalled.

We now ensure that authd is always added as a source to the `passwd`,
`group` and `shadow` databases in /etc/nsswitch.conf when the authd
package is installed.

That's done by always calling `insert_nss_entry` on install, which is
already implemented so that it only adds the `authd` string to those
`passwd|group|shadow` lines in /etc/nsswitch.conf which don't already
contain it, so we can just always call that function on install.

UDENG-3906