qrcodemodel: Rate limit the emission of reselect auth mode events
When hitting the Enter key for a while, bubbletea sends lots of events
and we may imply sending reselectAuthMode{} events faster than we can
actually handle the cancellation/reauthModeSelection in the whole stack.

This never happened (ever) when using the example broker, but it may
actually happen in a real world scenario where the broker may not cancel
the requests quickly enough.

Do this instead of avoiding the handling of the reselectAuthMode event
because it's better to prevent bubbletea to break our (fragile) system.