Rework dbusbroker dbus calls to format errors
We now have a wrapper around the dbus calls to be able to intercept the
errors and format the messages in a way that is helpful to end user.
This helps avoiding error messages that are too technical, such as "the
object X was not provided by any .service files" which is returned when
the broker is not available on DBus.

Co-authored-by: Didier Roche-Tolomelli <didrocks@ubuntu.com>