Avoid calling user.LookupId during authd startup
`user.LookupId` calls `getpwuid` which queries NSS for the UID. Since
authd is installed as a NSS source itself, this leads to a timeout when
authd hasn't started up yet, because it can't respond to the query.

We avoid that by using the UIDs to identify the active users instead of
the user names.