pam/nativemodel: Ensure current stage is always in sync with actual state
Due to the async nature of the models, we may end up not being in a
stage that the local current stage variable is tracking, because that
may be updated at later point, when the updateStage happens.

So instead of relying on the emission of nativeChangeStage, that may
happen even when the base model does it, define a native change state
request that also performs the required action after having changed the
state.