Replace usages of internal log package with log/slog
Except in the PAM module, which configures the log package in such a
complex way that it's out of scope of this commit. Since the PAM module
also configures the log package to use systemd journal's formatting when
printing to the journal, we might want to migrate it to log/slog once we
add systemd journal formatting for it (UDENG-4621).

The log messages from our internal log package used to look like this:

    time="2024-09-25T17:42:30+02:00" level=debug msg="Updating entry of user \"user1\" (UID: 1111)"

Now they look like this:

    2024/09/25 17:41:03 DEBUG Registered user "user1" with UID 1111