pam/integration-tests/native: Check that the user 'r' is accepted
This is something that was manually tested, but since 'r' is a special
string in the native model, let's add an explicit check, ensuring that
a such user can work properly.

And no, we won't support a 'r' secret, even though we could just wait
for a broker refusal before using it as a "go back" command.