Make UID/GID ranges configurable
It's hard to know for sure if a generated UID could be used by some
other user on the system (including systemd dynamic users, subuid(5),
etc.).

Instead of trying to do that, we make it the administrators
responsibility to ensure that the configured UID/GID ranges don't
overlap with the ranges used by other system components.

The default range (1000000000-1999999999) should not be used by anything
else by default.

UDENG-4781