pam/integration-tests: Add PAM tests using SSHd
We have tests simulating SSH behavior, but it's definitely better to
ensure that SSH works as expected using the actual server and client
when used with authd.

In order to get sshd to be fully usable for this simulation, however, we
need to "mock" it by using a LD_PRELOAD'ed library that has to be in C
(as the cgo version I initially done would trigger the well known issues
we have with go libraries and threads) and that we use it for mocking
the sshd requests on getpwnam and to make sshd to open our pam file
(that is hardcoded in sshd).

To handle the getpwnam we could even have used __nss_configure_lookup()
with a fake module or our own, but this is just a simpler solution for
now, while in future we may want to add full integration tests where
also our own NSS library is used instead, but this was outside the scope
of this change, that is mainly focused on the behavior of the PAM module
only.

As for the rest, just repeat all the native tests that make sense using
SSH instead, by de facto re-using the same tape files, minus the removal
of the user selection.