ci: Parallelize Go tests jobs
Tests start to take quite a bit of time an in case a specific test kind
fails we have to restart the whole CI job.

To avoid this use a matrix to parallelize the go test jobs and skip the
steps that are not required in each step.

This is not optimal since we still repeat the setup (although some parts
are cached and in general the common part takes less than a minute), but
it allows to get faster builds with a very minimal effort