localgroups: Avoid racy read/writes on local groups
We are relying on local groups content in order to call gpasswd, but
in case there are two concurrent writes, the gpasswd file content we're
reading may not be updated, causing us to call --add or --remove twice
for the same content.

So use a mutex to avoid that we're trying to read from the filesystem
before that the current write operation has been completed.