pam/integration-tests: Make tapes commands to be controllable from go side
We may want to replicate some tapes for multiple tests but with
different commands, so we need to define the command in a generic way
instead of repeating it in each single tape.

The variables are evaluated before running the tape, so no changes in
the golden files are needed.