pam/integration-tests: Use the same authd instance when possible
We were running multiple instances of authd daemon for each test that in
order to make some specific cases to be tested.
However this is unneeded for the great majority of the test cases and it
doesn't allow us to test the daemon concurrency properly.

So, just run a specific daemon if the test case requires it, and so in
just in case we need to do root-checks or to ensure that the local
groups are updated.

The only downside of this is that if a test generates a gpasswd file we
are going to fail also in other tests that are not affected by the issue
but that's still something that can be easily debugged checking the logs
or temporary enabling the single-authd instance to be used all the times