pam/integration-tests: Add tests using a single SSH server for all the requests
We may want to be sure that a single instance of SSH with multiple
requests coming in parallel is properly handled by our stack.

This is something we didn't test before but having sshd as a daemon
allows us to do it properly, simulating a more real scenario.

However, we only perform such tests in race mode not to increase the
testing time too much