pam/auhmodeselection: Do not send duplicated GetAuthenticationModes requests
In some cases we may end up sending to the broker two authentication
modes requests, which may lead to an unexpected authModeSelected for the
first mode on the 2nd reply of the method.

And so to a an unexpected challenge start.

See: https://github.com/ubuntu/authd/actions/runs/11325604248/job/31492851219

To prevent this, let's just not send any mode request on supported
layouts, since at this point we should already have them set already