pam/integration-tests/{ssh,native}: Use VHS Wait feature
Improves the reliability of tests by ensuring that we're sending input
only when something we're expecting is on screen.

Due to the fact that the native model doesn't clear the terminal, we
need to take care of checking that prompts that we expect are on screen.

We pin the version to the latest commit that is in main at the moment,
but we can safely go back to latest once a newer VHS will be released.