pam/formmodel: Use uniform style for password entries
In the CLI interface we used to use EchoPassword entry mode for the
new password form, while an empty field for other operations.

This is quite inconsistent, so since we've an unique type for handling
the input models (not used at the moment), just use it to ensure that
we're using always the same policy in all the password entries we have.