pam/integration-tests/gdm: Handle broker selection when on such stage
We have some races in gdm tests caused by the fact we may send the
brokerSelection event too late, in particular when we received the
broker list, but at that point we may be have done other steps already.

Instead, handle the broker selection only when we're in broker selection
stage, and not too early or too late.