pam/nativemodel: Error if broker does not provide a valid prompt label
We were trying to fix the broker behavior here, but it's something that
was done only by the native UI and it's not really something used
anywhere.

So instead just error in case this happens, since we expect to have
defined labels here.

In case we will ever require to have a fallback string, this should be
done at general level and not just at PAM level.