pam/model: Initialize user only after we've got the brokers list
One of the races we were facing (mostly in native model) is that we
could end up in a situation in which we receive an authd error on
getting the brokers list *after* that the user selection view is already
shown, leading to racy golden files ordering.

This is because what it could happen is:
 - We start in parallel:
   -> User selection
   -> UI Layouts -> Available brokers fetching

Now let's "imagine" that the brokers fetching fails (as when we've not
the permissions to read it, as for being not root).

If the user selection result arrives first, then we end up showing the
user selection UI (that is blocking in the native model) and only then
we handle the AvailableBrokers result (the error).

If the broker selection arrives first, we show first the error instead.

Now, this kind of approach made sense when we had no errors during this
phases, to get the user name as early as possible, but since we're now
doing various permissions checks in the broker, it's better to always go
in order instead of performing such operations in parallel, so now:
 - UI Layouts -> Available brokers fetching
 - User selection starts

Blocking as soon as we've a failure, as highlighted by the new golden
files.